Update Docs.tsx - Kareem read desc
fixed a few sections. The changes i made need to be changed across the entire site. Hopefully this gives you a little better understanding of the DSL and especially the workflow sections. 
also, change Workflow to Magnetic Field and you can say Magnetic Field (workflow) to clarify that magnetic fields are the workflow. 

Also you need to highlight teams, they are like the most foundational aspect of GLUE, without teams there is nothing. 

teams - agent to agent communication and collaboration within teams is not restricted like magentic field flows. they can communicate freely back and forth, 
the team lead - helps orchestrate the team communication and collaboration, sort of a traffic controller. 
also a HUGE selling point is that tools are given to TEAMS, rather than a specific individual agent, this allows agents to use multiple tools within the same session, 

models are given adhesives which effect tool output persistence and if its shared with the team. 

adhesives have nothing to do with directing the flow of information. that is ALL the magnetic field. 

>< bidirectional -> push <> repel <- pull (a fallback, only one team can be the pull team, and its declared exactly like this: 

team1 >< team2
team2 -> team3
team4 <- pull       - literally just "pull" with the <- operator declares that team to be able to use the pull as a fallback 


also highlight that no matter what teams are used, even if your app only has one agent. or if a team has one agent, then its naturally the lead, but that doesnt need to be declared if there are no other members, the team lead only needs to be declared if there are more than one member